requirements.txt
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.3.2
shap==0.44.1
streamlit==1.29.0
streamlit-option-menu==0.3.6
matplotlib==3.8.2
seaborn>=0.12.0
Pillow>=10.0.0
supabase>=2.3.0
python-dotenv>=1.0.0
